test: split the tests into files based on functions
